moving environment service common code to abstract class
