+++ src/services/environments/node-environment.service.ts
@@ -1,16 +1,16 @@
 import * as _path from 'path';
-import { exec, spawn } from 'child_process';
+import { exec } from 'child_process';
 
-import { IEnvironmentService } from "./environment.service";
+import { EnvironmentServiceBase } from "./environment.service";
 import { Solution } from '../../models/solution';
 import { FileService, IFileService } from '../file.service';
 
-export class NodeEnvironmentService implements IEnvironmentService {
+export class NodeEnvironmentService extends EnvironmentServiceBase {
 
   fileService: IFileService = new FileService();
   /* istanbul ignore next */
   private npm: string = process.platform === 'win32' ? 'npm.cmd' : 'npm';
-
+  
   updateProjectDefinition = (folderPath: string, projectName: string, solution: Solution): Promise<void> => {
     let packagePath = _path.join(folderPath, 'package.json');
     return this.fileService.readJson<any>(packagePath).then(pkg => {
@@ -26,29 +26,6 @@ export class NodeEnvironmentService implements IEnvironmentService {
       }
       return this.fileService.writeJson(packagePath, pkg);
     });
-  }
-
-  addProjectScaffoldFile = (folderPath: string, projectName: string, solution: Solution): Promise<void> => {
-    let project = solution?.projects.find(p => p.name == projectName);
-    const scaffoldFile = {
-      name: projectName,
-      include: project?.include ?? [],
-      specs: project?.specs ?? {}
-    }
-    let scaffoldFilePath = _path.join(folderPath, 'shaman.scaffold.json');
-    return this.fileService.writeJson(scaffoldFilePath, scaffoldFile);
-  }
-
-  executeProjectScaffolding = (folderPath: string): Promise<void> => {
-    let scaffoldCommand = new Promise((res, err) => {
-      let childProcess = spawn('node', ['shaman.scaffold.js'], {cwd: folderPath});
-      childProcess.stdout.on('data', (data) => process.stdout.write(`${data}`));
-      childProcess.stderr.on('data', (data) => process.stderr.write(`${data}`));
-      childProcess.on('close', (code) => code === 0 ? res(null) : err(
-        new Error("An error occurred while executing project scaffolding.")
-      ));
-    });
-    return scaffoldCommand.then(_ => this.fileService.deleteFile(_path.join(folderPath, 'shaman.scaffold.js')));
   }
 
   installDependencies = (folderPath: string, projectName: string): Promise<void> => {
